src/requirements.txt
aiobreaker==1.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
asyncio==4.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
drf-orjson-renderer==1.7.3
drf-spectacular==0.28.0
frozenlist==1.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
multidict==6.6.4
orjson==3.11.3
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.22
pyOpenSSL==25.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
tenacity==9.1.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
uritemplate==4.2.0
yarl==1.20.1
zope.interface==7.2
